backend/requirements.txt
Django==3.2.3
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
requests==2.31.0
Pillow==9.3.0